pyproject.toml
[tool.poetry]
name = "public-ip"
version = "0.1.0"
description = ""
authors = ["meirdev <maornet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
dnspython = "^2.2.1"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
mypy = "^0.982"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
